Integracion/requirements.txt
pandas==1.5.3
psycopg2==2.9.10
psycopg2_binary==2.9.10
python-dotenv==1.2.1
SQLAlchemy==2.0.45
